requirements.txt
django==2.1
correios==4.7.1
django_heroku==0.3.1
django-bootstrap-form==3.4
lxml==4.2.1
requests
zeep==3.1.0
django-bootstrap-modal-forms==1.3.1